requirements.txt
click==8.0.1
python-dotenv==0.19.0
pytest==6.2.4
flake8==3.9.2
